style(bb-test): split aggregate evaluation and satisfy changed-style gate
