+++ packages/shared-test/rallar-bb-test/distributed/group-assertions-evaluation.ts
@@ -1,12 +1,10 @@
 // deno-lint-ignore-file no-explicit-any
-import { assertValueMatches } from '../assert/assert-value-operators.ts';
 import { redactRallarBlackBoxValue } from '../redaction.ts';
 import type { RallarBlackBoxTestRedactionOptions } from '../types.ts';
 import type {
     RallarBlackBoxDistributedRecipeResult,
     RallarBlackBoxDistributedRunItemState,
     RallarBlackBoxDistributedRunManifest,
-    RallarBlackBoxDistributedTargetResolution,
 } from '../distributed-run.ts';
 import {
     RALLAR_BB_DISTRIBUTED_GROUP_ASSERTION_EVIDENCE_MISSING,
@@ -16,6 +14,10 @@ import {
     type RallarBlackBoxDistributedGroupAssertionResult,
     type RallarBlackBoxGroupAssertionAgentRow,
 } from './group-assertions.ts';
+import {
+    evaluateGroupAssertionAggregate,
+    type GroupAssertionVerdict,
+} from './group-assertions-aggregates.ts';
 import {
     collectGroupAssertionEvidence,
     type DistributedGroupAssertionParticipant,
@@ -32,30 +34,6 @@ const COMPLETED_RECIPE_STATES: readonly RallarBlackBoxDistributedRunItemState[] 
     'skipped',
 ];
 
-// Group agreement equality: object key order is irrelevant, array order is
-// significant. Deliberately distinct from sameJsonValue (stringify equality)
-// and from json-compare's order-insensitive exact mode.
-export function deepEqualJson(left: any, right: any): boolean {
-    if (left === right) {
-        return true;
-    }
-    if (left === null || right === null || typeof left !== 'object' || typeof right !== 'object') {
-        return false;
-    }
-    if (Array.isArray(left) || Array.isArray(right)) {
-        return Array.isArray(left) &&
-            Array.isArray(right) &&
-            left.length === right.length &&
-            left.every((entry, index) => deepEqualJson(entry, right[index]));
-    }
-    const leftKeys = Object.keys(left);
-    return leftKeys.length === Object.keys(right).length &&
-        leftKeys.every(key =>
-            Object.prototype.hasOwnProperty.call(right, key) &&
-            deepEqualJson(left[key], right[key])
-        );
-}
-
 export interface EvaluateDistributedGroupAssertionsInput {
     readonly manifest: RallarBlackBoxDistributedRunManifest;
     readonly participants: readonly DistributedGroupAssertionParticipant[];
@@ -65,7 +43,7 @@ export interface EvaluateDistributedGroupAssertionsInput {
 }
 
 // Coordinator-side evaluation over the frozen participant set. Returns
-// undefined while required recipes are still executing; the control-server
+// undefined while dispatched recipes are still executing; the control-server
 // rollup calls this on every refresh, so the first fully completed pass is
 // the one that decides.
 export function evaluateDistributedGroupAssertions(
@@ -88,11 +66,10 @@ function evaluateGroupAssertion(
     assertion: RallarBlackBoxDistributedGroupAssertion,
     input: EvaluateDistributedGroupAssertionsInput,
 ): RallarBlackBoxDistributedGroupAssertionResult {
-    const scopedParticipants = assertion.scope?.role === undefined
+    const scopeRole = assertion.scope?.role;
+    const scopedParticipants = scopeRole === undefined
         ? input.participants
-        : input.participants.filter(participant =>
-            participant.roles.includes(assertion.scope!.role)
-        );
+        : input.participants.filter(participant => participant.roles.includes(scopeRole));
     if (scopedParticipants.length === 0) {
         return toNoParticipantsResult(assertion);
     }
@@ -140,149 +117,6 @@ function evaluateGroupAssertion(
     };
 }
 
-interface GroupAssertionVerdict {
-    readonly ok: boolean;
-    readonly violatingAgentIds: readonly string[];
-    readonly matchingCount?: number;
-    readonly reason: string;
-    readonly detail?: any;
-    readonly verdictByAgentId: ReadonlyMap<string, RallarBlackBoxGroupAssertionAgentRow['verdict']>;
-}
-
-function evaluateGroupAssertionAggregate(
-    assertion: RallarBlackBoxDistributedGroupAssertion,
-    resolved: readonly GroupAssertionEvidenceRow[],
-): GroupAssertionVerdict {
-    switch (assertion.aggregate) {
-        case 'allMatch':
-        case 'noneMatch':
-        case 'countMatching':
-            return evaluatePredicateAggregate(assertion, resolved);
-        case 'allEqual':
-            return evaluateAllEqual(resolved);
-        case 'allEqualWithin':
-            return evaluateAllEqualWithin(resolved, assertion.tolerance);
-    }
-}
-
-function evaluatePredicateAggregate(
-    assertion: RallarBlackBoxDistributedGroupAssertion & Readonly<{
-        aggregate: 'allMatch' | 'noneMatch' | 'countMatching';
-    }>,
-    resolved: readonly GroupAssertionEvidenceRow[],
-): GroupAssertionVerdict {
-    const verdictByAgentId = new Map<string, RallarBlackBoxGroupAssertionAgentRow['verdict']>();
-    const matchingAgentIds: string[] = [];
-    for (const row of resolved) {
-        const matches = assertValueMatches(
-            { exists: true, value: row.value },
-            assertion.predicate.operator,
-            assertion.predicate.expected,
-        );
-        verdictByAgentId.set(row.agentId, matches ? 'matching' : 'not-matching');
-        if (matches) {
-            matchingAgentIds.push(row.agentId);
-        }
-    }
-    const matchingCount = matchingAgentIds.length;
-
-    if (assertion.aggregate === 'allMatch') {
-        const violating = resolved
-            .map(row => row.agentId)
-            .filter(agentId => !matchingAgentIds.includes(agentId));
-        return {
-            ok: violating.length === 0,
-            violatingAgentIds: violating,
-            matchingCount,
-            reason: `${matchingCount} of ${resolved.length} participants matched the predicate`,
-            verdictByAgentId,
-        };
-    }
-    if (assertion.aggregate === 'noneMatch') {
-        return {
-            ok: matchingCount === 0,
-            violatingAgentIds: matchingAgentIds,
-            matchingCount,
-            reason: `${matchingCount} participants matched a predicate none may match`,
-            verdictByAgentId,
-        };
-    }
-    const bounds = assertion.count;
-    const ok = (bounds.equals === undefined || matchingCount === bounds.equals) &&
-        (bounds.gte === undefined || matchingCount >= bounds.gte) &&
-        (bounds.lte === undefined || matchingCount <= bounds.lte);
-    return {
-        ok,
-        violatingAgentIds: [],
-        matchingCount,
-        reason: `matching count ${matchingCount} violates ${JSON.stringify(bounds)}`,
-        detail: { count: bounds },
-        verdictByAgentId,
-    };
-}
-
-function evaluateAllEqual(resolved: readonly GroupAssertionEvidenceRow[]): GroupAssertionVerdict {
-    const classes: { value: any; agentIds: string[] }[] = [];
-    for (const row of resolved) {
-        const existing = classes.find(candidate => deepEqualJson(candidate.value, row.value));
-        if (existing) {
-            existing.agentIds.push(row.agentId);
-        } else {
-            classes.push({ value: row.value, agentIds: [row.agentId] });
-        }
-    }
-    const reference = [...classes].sort((left, right) =>
-        right.agentIds.length - left.agentIds.length ||
-        smallestAgentId(left).localeCompare(smallestAgentId(right))
-    )[0];
-    const violating = classes
-        .filter(candidate => candidate !== reference)
-        .flatMap(candidate => candidate.agentIds);
-    const verdictByAgentId = new Map<string, RallarBlackBoxGroupAssertionAgentRow['verdict']>();
-    for (const row of resolved) {
-        verdictByAgentId.set(row.agentId, violating.includes(row.agentId) ? 'violating' : 'agreeing');
-    }
-    return {
-        ok: classes.length <= 1,
-        violatingAgentIds: violating,
-        reason: `${classes.length} distinct values across ${resolved.length} participants`,
-        detail: reference === undefined ? undefined : { referenceValue: reference.value },
-        verdictByAgentId,
-    };
-}
-
-function evaluateAllEqualWithin(
-    resolved: readonly GroupAssertionEvidenceRow[],
-    tolerance: number,
-): GroupAssertionVerdict {
-    const nonNumeric = resolved.filter(row => typeof row.value !== 'number');
-    const numericRows = resolved.filter(row => typeof row.value === 'number');
-    const values = numericRows.map(row => row.value as number);
-    const minValue = values.length === 0 ? undefined : Math.min(...values);
-    const maxValue = values.length === 0 ? undefined : Math.max(...values);
-    const spread = minValue === undefined || maxValue === undefined
-        ? 0
-        : maxValue - minValue;
-    const withinTolerance = spread <= tolerance;
-    const extremeAgentIds = withinTolerance ? [] : numericRows
-        .filter(row => row.value === minValue || row.value === maxValue)
-        .map(row => row.agentId);
-    const violating = [...nonNumeric.map(row => row.agentId), ...extremeAgentIds];
-    const verdictByAgentId = new Map<string, RallarBlackBoxGroupAssertionAgentRow['verdict']>();
-    for (const row of resolved) {
-        verdictByAgentId.set(row.agentId, violating.includes(row.agentId) ? 'violating' : 'agreeing');
-    }
-    return {
-        ok: nonNumeric.length === 0 && withinTolerance,
-        violatingAgentIds: violating,
-        reason: nonNumeric.length > 0
-            ? `${nonNumeric.length} participants reported non-numeric values`
-            : `numeric spread ${spread} exceeds tolerance ${tolerance}`,
-        detail: { minValue, maxValue, spread, tolerance },
-        verdictByAgentId,
-    };
-}
-
 function toRedactedAgentRows(
     rows: readonly GroupAssertionEvidenceRow[],
     verdict: GroupAssertionVerdict,
@@ -292,18 +126,11 @@ function toRedactedAgentRows(
         agentId: row.agentId,
         role: row.role,
         evidence: row.status,
-        verdict: verdictByAgent(verdict, row),
+        verdict: row.status === 'resolved' ? verdict.verdictByAgentId.get(row.agentId) : undefined,
         value: row.status === 'resolved'
             ? redactRallarBlackBoxValue(row.value, redaction)
             : undefined,
     }));
-}
-
-function verdictByAgent(
-    verdict: GroupAssertionVerdict,
-    row: GroupAssertionEvidenceRow,
-): RallarBlackBoxGroupAssertionAgentRow['verdict'] {
-    return row.status === 'resolved' ? verdict.verdictByAgentId.get(row.agentId) : undefined;
 }
 
 interface ToGroupAssertionErrorInput {
@@ -369,8 +196,4 @@ function toNoParticipantsResult(
             message: `Group assertion ${assertion.groupAssertionId} cannot evaluate: ${scopeText}.`,
         },
     };
-}
-
-function smallestAgentId(candidate: Readonly<{ agentIds: readonly string[] }>): string {
-    return [...candidate.agentIds].sort((left, right) => left.localeCompare(right))[0] ?? '';
 }
